test(sb3): Share the transaction-monitoring cached context with the endpoint test
Fold TransactionMonitoringEndpointTest into the shared
AbstractTransactionMonitoringSharedContextTest context so the endpoint test
hits the Spring TestContext cache instead of building its own context.

The parent now owns the union of all configuration and fixtures (RANDOM_PORT
web env, the management-endpoint property source, JwtAuthTestConfiguration, the
endpoint/service/appender beans, and the richer Owner/Pet/OwnerRepository/
PropagationTestHelper) and never references its subclasses. The endpoint test
is reduced to extending the parent.

spring-test-profiler confirms all three transaction test classes
(BeanPostProcessor, ProxyingDataSource, Endpoint) now resolve to a single
cached ApplicationContext.
