+++ sbs/axelix-spring-boot-3-starter/src/test/java/com/axelixlabs/axelix/sbs/spring/core/transactions/AbstractTransactionMonitoringSharedContextTest.java
@@ -17,27 +17,42 @@
  */
 package com.axelixlabs.axelix.sbs.spring.core.transactions;
 
+import java.util.ArrayList;
 import java.util.List;
 
+import jakarta.persistence.CascadeType;
 import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.Table;
 
 import io.micrometer.core.instrument.MeterRegistry;
 
 import org.springframework.beans.factory.ObjectProvider;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.boot.context.event.ApplicationReadyEvent;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Import;
 import org.springframework.context.annotation.Lazy;
+import org.springframework.context.event.EventListener;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+import org.springframework.test.context.TestPropertySource;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.axelixlabs.axelix.sbs.spring.core.auth.JwtAuthTestConfiguration;
 import com.axelixlabs.axelix.sbs.spring.core.metrics.AxelixMetricsPublisher;
 import com.axelixlabs.axelix.sbs.spring.core.metrics.DefaultAxelixMetricsPublisher;
 
@@ -47,19 +62,41 @@ import com.axelixlabs.axelix.sbs.spring.core.metrics.DefaultAxelixMetricsPublish
  * so that all subclasses resolve to an identical merged context configuration and therefore share a
  * single cached {@link org.springframework.context.ApplicationContext}.
  *
+ * <p>The context runs with a {@link SpringBootTest.WebEnvironment#RANDOM_PORT} web server and exposes the
+ * transaction-monitoring actuator endpoint, so the endpoint integration test shares the very same cached
+ * context as the non-web bean-post-processor tests. <strong>All</strong> test configurations and fixtures
+ * consumed by the transaction-monitoring tests live here, in the parent, as nested classes — so the parent
+ * owns the entire shared configuration and never has to reference its subclasses.
+ *
  * @author Sergey Cherkasov
  * @author Nikita Kirillov
  * @author Artemiy Degtyarev
  * @author Mikhail Polivakha
  */
-@SpringBootTest
-@Import(AbstractTransactionMonitoringSharedContextTest.SharedTransactionTestConfiguration.class)
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@TestPropertySource(properties = {"management.endpoints.web.exposure.include=axelix-transactions-monitoring"})
+@Import({
+    AbstractTransactionMonitoringSharedContextTest.SharedTransactionTestConfiguration.class,
+    JwtAuthTestConfiguration.class
+})
 abstract class AbstractTransactionMonitoringSharedContextTest {
 
     @TestConfiguration
     @EnableJpaRepositories(basePackageClasses = OwnerRepository.class, considerNestedRepositories = true)
-    @EntityScan(basePackageClasses = Owner.class)
+    @EntityScan(basePackageClasses = {Owner.class, Pet.class})
     static class SharedTransactionTestConfiguration {
+
+        @Bean
+        public TransactionMonitoringEndpoint transactionMonitoringEndpoint(
+                TransactionMonitoringService transactionMonitoringService) {
+            return new TransactionMonitoringEndpoint(transactionMonitoringService);
+        }
+
+        @Bean
+        public TransactionMonitoringService transactionMonitoringService(
+                TransactionStatsCollector transactionStatsCollector) {
+            return new DefaultTransactionMonitoringService(transactionStatsCollector);
+        }
 
         @Bean
         public TransactionStatsCollector transactionStatsCollector() {
@@ -102,9 +139,15 @@ abstract class AbstractTransactionMonitoringSharedContextTest {
         public AxelixMetricsPublisher axelixMetricsPublisher(MeterRegistry meterRegistry) {
             return new DefaultAxelixMetricsPublisher(meterRegistry);
         }
+
+        @EventListener(ApplicationReadyEvent.class)
+        public void registerAppender() {
+            new LogbackInMemoryPaginationAppenderRegistrar().register();
+        }
     }
 
     @Entity
+    @Table(name = "owner")
     static class Owner {
 
         @Id
@@ -113,6 +156,13 @@ abstract class AbstractTransactionMonitoringSharedContextTest {
 
         private String lastName;
 
+        @OneToMany(mappedBy = "owner", cascade = CascadeType.PERSIST, fetch = FetchType.LAZY)
+        private List<Pet> pets = new ArrayList<>();
+
+        public List<Pet> getPets() {
+            return pets;
+        }
+
         public Long getId() {
             return id;
         }
@@ -120,19 +170,55 @@ abstract class AbstractTransactionMonitoringSharedContextTest {
         public String getLastName() {
             return lastName;
         }
+
+        public Owner setLastName(String lastName) {
+            this.lastName = lastName;
+            return this;
+        }
+
+        public Owner addPet(Pet pet) {
+            this.pets.add(pet);
+            return this;
+        }
+    }
+
+    @Entity
+    @Table(name = "pet")
+    static class Pet {
+
+        @Id
+        @GeneratedValue(strategy = GenerationType.IDENTITY)
+        private Long id;
+
+        private String name;
+
+        @ManyToOne
+        @JoinColumn(name = "owner_id")
+        private Owner owner;
+
+        public Pet() {}
+
+        public Pet(String name, Owner owner) {
+            this.name = name;
+            this.owner = owner;
+        }
     }
 
     interface OwnerRepository extends JpaRepository<Owner, Long> {
 
         @Transactional
-        default Owner findByLastName(String lastName) {
-            return new Owner();
-        }
+        Owner findByLastName(String lastName);
 
         @Transactional(propagation = Propagation.SUPPORTS)
         default List<Owner> findAll() {
             return List.of(new Owner());
         }
+
+        @Transactional
+        @Query(
+                value = "SELECT o FROM AbstractTransactionMonitoringSharedContextTest$Owner o JOIN FETCH o.pets",
+                countQuery = "SELECT COUNT(o) FROM AbstractTransactionMonitoringSharedContextTest$Owner o")
+        Page<Owner> findAllWithPets(Pageable pageable);
     }
 
     static class PropagationTestHelper {
@@ -143,6 +229,62 @@ abstract class AbstractTransactionMonitoringSharedContextTest {
         public PropagationTestHelper(OwnerRepository ownerRepository, @Lazy PropagationTestHelper self) {
             this.ownerRepository = ownerRepository;
             this.self = self;
+        }
+
+        // IMPORTANT: Calling via 'self' proxy is required to properly test the REQUIRED -> REQUIRES_NEW stack behavior.
+        @Transactional(propagation = Propagation.REQUIRED)
+        public void outerRequiredMethod(String outerName) {
+            ownerRepository.save(new Owner().setLastName(outerName));
+
+            self.saveRequiresNew("SomeName");
+
+            ownerRepository.save(new Owner().setLastName(outerName));
+        }
+
+        @Transactional(propagation = Propagation.REQUIRES_NEW)
+        public void saveRequiresNew(String lastName) {
+            ownerRepository.save(new Owner().setLastName(lastName));
+        }
+
+        @Transactional(propagation = Propagation.REQUIRES_NEW)
+        public void testSaveMultipleOwners() {
+            ownerRepository.saveAll(List.of(new Owner(), new Owner(), new Owner()));
+        }
+
+        @Transactional(propagation = Propagation.REQUIRES_NEW)
+        public void updateOwner(Owner owner) {
+            // will cause entityManager.merge --> new SELECT, since Owner has an id
+            ownerRepository.save(owner);
+        }
+
+        @Transactional(propagation = Propagation.REQUIRES_NEW)
+        public void findOwnerById(Long id) {
+            Owner owner = ownerRepository.findById(id).orElseThrow();
+            owner.getPets().size(); // will cause n + 1
+        }
+
+        @Transactional(propagation = Propagation.NESTED)
+        public void testNested() {
+            ownerRepository.findByLastName("Schroeder");
+        }
+
+        @Transactional(propagation = Propagation.SUPPORTS)
+        public void testSupports(String lastName) {
+            ownerRepository.findByLastName(lastName);
+        }
+
+        @Transactional(propagation = Propagation.SUPPORTS)
+        public void testSupportsWithoutTransaction() {}
+
+        @Transactional
+        public void testRollbackScenario(String lastName) {
+            ownerRepository.findByLastName(lastName);
+            throw new RuntimeException("Test rollback");
+        }
+
+        @Transactional(propagation = Propagation.REQUIRES_NEW)
+        public void findAllWithPetsPageable() {
+            ownerRepository.findAllWithPets(PageRequest.of(0, 5));
         }
 
         @Transactional(propagation = Propagation.REQUIRES_NEW)
